+++ odoo/local-src/connector_qoqa/__init__.py
@@ -20,11 +20,9 @@ from . import product_attribute
 from . import product_attribute_value
 from . import promo
 from . import qoqa_backend
-from . import qoqa_discount_accounting
 from . import qoqa_offer
 from . import qoqa_shipping_fee
 from . import qoqa_shop
-from . import qoqa_voucher_payment
 from . import res_company
 from . import res_country
 from . import res_currency