+++ odoo/local-src/connector_qoqa/__openerp__.py
@@ -3,7 +3,7 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html)
 
 {'name': 'QoQa Connector',
- 'version': '9.0.2.0.0',
+ 'version': '9.0.2.1.0',
  'category': 'Connector',
  'depends': ['connector',
              'crm_claim',