+++ odoo/local-src/connector_qoqa/sale_line/common.py
@@ -51,3 +51,4 @@ class SaleOrderLine(models.Model):
         string='QBindings',
         context={'active_test': False},
     )
+    is_voucher = fields.Boolean(string='Is a Voucher')