+++ odoo/local-src/connector_qoqa/tests/test_import_order.py
@@ -22,7 +22,10 @@ ExpectedOrder = namedtuple(
 )
 ExpectedOrderLine = namedtuple(
     'ExpectedOrderLine',
-    'product_id name price_unit product_uom_qty'
+    'product_id name price_unit product_uom_qty is_voucher'
+)
+ExpectedOrderLine.__new__.__defaults__ = (
+    (False,) * len(ExpectedOrderLine._fields)
 )
 
 
@@ -38,6 +41,7 @@ class TestImportOrder(QoQaTransactionCase):
 
     def setup_order_dependencies(self):
         method = self.env.ref('account.account_payment_method_manual_in')
+        self.payment_method = method
         self.payment_mode = self.env['account.payment.mode'].create({
             'name': 'Paypal',
             'payment_method_id': method.id,
@@ -235,6 +239,49 @@ class TestImportOrder(QoQaTransactionCase):
                 name='Bon marketing (10000011)',
                 price_unit=-499,
                 product_uom_qty=1,
+            ),
+        ]
+        self.assert_records(expected, order.order_line)
+
+    @freeze_time('2016-04-28 00:00:00')
+    @recorder.use_cassette()
+    def test_import_sale_order_with_voucher(self):
+        voucher_product = self.env['product.product'].create({
+            'name': 'Bon Cadeau',
+            'default_code': 'BC',
+        })
+        self.backend_record.voucher_product_id = voucher_product
+        self.env['account.payment.mode'].create({
+            'name': 'Bon Cadeau',
+            'payment_method_id': self.payment_method.id,
+            'company_id': self.env.ref('base.main_company').id,
+            'bank_account_link': 'variable',
+            'qoqa_id': '9',
+        })
+        import_record(self.session, 'qoqa.sale.order',
+                      self.backend_record.id, 4260998)
+        domain = [('qoqa_id', '=', '4260998')]
+        order = self.env['qoqa.sale.order'].search(domain)
+        order.ensure_one()
+        expected = [
+            ExpectedOrderLine(
+                product_id=self.product_1_binding.openerp_id,
+                name='[product_1] Product 1',
+                price_unit=79.,
+                product_uom_qty=1,
+            ),
+            ExpectedOrderLine(
+                product_id=self.drone_product,
+                name='Drone delivery',
+                price_unit=9,
+                product_uom_qty=1,
+            ),
+            ExpectedOrderLine(
+                product_id=voucher_product,
+                name='Bon Cadeau (562614)',
+                price_unit=-50.,
+                product_uom_qty=1.,
+                is_voucher=True,
             ),
         ]
         self.assert_records(expected, order.order_line)